add event for channel and cleanup
