+++ src/main/java/org/openhab/automation/jrule/internal/handler/JRuleEventHandler.java
@@ -1,13 +1,13 @@
 /**
  * Copyright (c) 2010-2022 Contributors to the openHAB project
- *
+ * <p>
  * See the NOTICE file(s) distributed with this work for additional
  * information.
- *
+ * <p>
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * http://www.eclipse.org/legal/epl-2.0
- *
+ * <p>
  * SPDX-License-Identifier: EPL-2.0
  */
 package org.openhab.automation.jrule.internal.handler;
@@ -588,4 +588,5 @@ public class JRuleEventHandler {
     private void logError(String message, Object... parameters) {
         JRuleLog.error(logger, LOG_NAME_EVENT, message, parameters);
     }
+
 }
